pyproject.toml
[tool.poetry]
name = "papermint"
version = "0.1.0"
description = ""
authors = ["goldenglorys <olusolagloryolamide@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.38.0"
pandoc = "^2.4"
jq = "^1.8.0"
requests = "^2.32.3"
pillow = "^10.4.0"
pypdf2 = "^3.0.1"
pylatex = "^1.4.2"
arxiv = "^2.1.3"
pdflatex = "^0.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
